+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Other
 
 * Avoid allocation in `World.Has(entity, compID)` (#16)
+* `World.RemEntity(entity)` panics on dead entity, like all other `World` methods (#18)
 
 ## [[v0.1.0]](https://github.com/mlange-42/arche/tree/v0.1.0)